+++ src/Plainion.Windows/Controls/Text/FileSystemDocumentStore.cs
@@ -173,7 +173,7 @@ namespace Plainion.Windows.Controls.Text
 
             return myIndex.Root.Enumerate()
                 .OfType<Document>()
-                .Where(doc => DocumentOperations.Search(doc.Body, doc.Body.ContentStart, text, SearchMode.Initial).Any())
+                .Where(doc => DocumentOperations.Search(doc.Body.Content(), doc.Body.ContentStart, text, SearchMode.Initial).Any())
                 .ToList();
         }